Reorganize MOI wrapper: split up into multiple files.
