+++ src/moi/nonlinear_constraints.jl
@@ -1,0 +1,6 @@
+MOI.supports(::Optimizer, ::MOI.NLPBlock) = true
+
+function MOI.set(model::Optimizer, ::MOI.NLPBlock, nlp_block_data::MOI.NLPBlockData)
+    model.nlp_block_data = nlp_block_data
+    return
+end
