+++ test/moi_test.jl
@@ -1,4 +1,7 @@
+module MOITests
+
 using BARON
+using Test
 
 using MathOptInterface
 const MOI = MathOptInterface
@@ -12,10 +15,12 @@ const config = MOIT.TestConfig(atol=1e-6, rtol=1e-6)
     MOIT.contlineartest(MOIB.SplitInterval{Float64}(optimizer), config)
 end
 
-@testset "MOI Integer Linear" begin
-    MOIT.intlineartest(MOIB.SplitInterval{Float64}(optimizer), config)
-end
+# @testset "MOI Integer Linear" begin
+#     MOIT.intlineartest(MOIB.SplitInterval{Float64}(optimizer), config)
+# end
 
-@testset "MOI Nonlinear" begin
-    MOIT.nonlineartest(MOIB.SplitInterval{Float64}(optimizer), config)
-end
+# @testset "MOI Nonlinear" begin
+#     MOIT.nonlineartest(MOIB.SplitInterval{Float64}(optimizer), config)
+# end
+
+end # module
