+++ test/runtests.jl
@@ -1,8 +1,8 @@
-include("gear.jl")
-include("minlp.jl")
-include("nlp1.jl")
-include("nlp2.jl")
-include("pool1.jl")
-include("bad_expressions.jl")
+# include("gear.jl")
+# include("minlp.jl")
+# include("nlp1.jl")
+# include("nlp2.jl")
+# include("pool1.jl")
+# include("bad_expressions.jl")
 
 include("moi_test.jl")
